% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacman.R
\name{pacman}
\alias{pacman}
\title{Package Loading and Installation Function}
\usage{
pacman(name, repolist = NULL)
}
\arguments{
\item{name}{vector of packages to be loaded, packages should be in quotationmarks}

\item{repolist}{optional vector of repositories (URLS in quotation marks) to install packages from if CRAN is not useful. If used, this needs to be ordered: URLS need to be in the same position of the repolist vector as the respective packages in the name vector. All packages desired to be downloaded from CRAN should be labelled "CRAN" in the repolist vector. If all packages are to be downloaded from CRAN, the repolist option can be left out.}
}
\description{
Checks whether Packages to be loaded have already been installed.
If yes: loads packages
If not: installs packages and loads them afterwards
}
